requirements.txt
aerich==0.6.3
asyncpg==0.25.0
fastapi==0.78.0
tortoise-orm==0.19.1
uvicorn==0.17.6
